Adds start minimized to system tray option
Implements a new setting that allows the application to start minimized in the system tray.

This change introduces a 'Start minimized to system tray' option in the general settings. When enabled, the application will start minimized and hide the main window, improving the user experience by reducing initial window clutter.

It also checks for the presence of an API key on startup. If an API key is not set and the "start minimized" setting is enabled, the application will show the settings page with the API config section to prompt the user to set up their API key.